+++ android/rctmgl/src/main/java/com/mapbox/rctmgl/components/styles/RCTMGLStyle.java
@@ -8,6 +8,7 @@ import com.facebook.react.bridge.ReadableMap;
 import com.facebook.react.bridge.ReadableMapKeySetIterator;
 import com.mapbox.mapboxsdk.maps.MapboxMap;
 import com.mapbox.rctmgl.utils.DownloadMapImageTask;
+import com.mapbox.rctmgl.utils.ImageEntry;
 
 import java.util.AbstractMap;
 import java.util.ArrayList;
@@ -63,6 +64,10 @@ public class RCTMGLStyle {
         addImage(styleValue, null);
     }
 
+    public ImageEntry imageEntry(RCTMGLStyleValue styleValue) {
+        return new ImageEntry(styleValue.getImageURI(), styleValue.getImageScale());
+    }
+
     public void addImage(RCTMGLStyleValue styleValue, DownloadMapImageTask.OnAllImagesLoaded callback) {
         if (!styleValue.shouldAddImage()) {
             if (callback != null) {
@@ -72,7 +77,7 @@ public class RCTMGLStyle {
         }
 
         String uriStr = styleValue.getImageURI();
-        Map.Entry[] images = new Map.Entry[]{ new AbstractMap.SimpleEntry(uriStr, uriStr) };
+        Map.Entry[] images = new Map.Entry[]{ new AbstractMap.SimpleEntry(uriStr, imageEntry(styleValue)) };
         DownloadMapImageTask task = new DownloadMapImageTask(mContext, mMap, callback);
         task.execute(images);
     }