+++ android/rctmgl/src/main/java/com/mapbox/rctmgl/utils/ConvertUtils.java
@@ -8,6 +8,8 @@ import com.facebook.react.bridge.Arguments;
 import com.facebook.react.bridge.NoSuchKeyException;
 import com.facebook.react.bridge.ReadableArray;
 import com.facebook.react.bridge.ReadableMap;
+import com.facebook.react.bridge.ReadableMapKeySetIterator;
+import com.facebook.react.bridge.ReadableType;
 import com.facebook.react.bridge.WritableArray;
 import com.facebook.react.bridge.WritableMap;
 import com.google.gson.JsonArray;
@@ -27,6 +29,104 @@ import java.util.Map;
 
 public class ConvertUtils {
     public static final String LOG_TAG = ConvertUtils.class.getSimpleName();
+
+    public static JsonObject toJsonObject(ReadableMap map) {
+        if (map == null) return null;
+        JsonObject result = new JsonObject();
+        ReadableMapKeySetIterator it = map.keySetIterator();
+
+        while (it.hasNextKey()) {
+            String key = it.nextKey();
+            switch (map.getType(key)) {
+                case Map:
+                    result.add(key, toJsonObject(map.getMap(key)));
+                    break;
+                case Array:
+                    result.add(key, toJsonArray(map.getArray(key)));
+                    break;
+                case Null:
+                    result.add(key, null);
+                    break;
+                case Number:
+                    result.addProperty(key, map.getDouble(key));
+                    break;
+                case String:
+                    result.addProperty(key, map.getString(key));
+                    break;
+                case Boolean:
+                    result.addProperty(key, map.getBoolean(key));
+                    break;
+            }
+        }
+        return result;
+    }
+
+    public static JsonArray toJsonArray(ReadableArray array) {
+        if (array == null) return null;
+        JsonArray result = new JsonArray(array.size());
+        for (int i = 0; i < array.size(); i++) {
+            switch (array.getType(i)) {
+                case Map:
+                    result.add(toJsonObject(array.getMap(i)));
+                    break;
+                case Array:
+                    result.add(toJsonArray(array.getArray(i)));
+                    break;
+                case Null:
+                    result.add((JsonElement)null);
+                    break;
+                case Number:
+                    result.add(array.getDouble(i));
+                    break;
+                case String:
+                    result.add(array.getString(i));
+                    break;
+                case Boolean:
+                    result.add(array.getBoolean(i));
+                    break;
+            }
+        }
+        return result;
+    }
+
+    public static JsonElement typedToJsonElement(ReadableMap map) {
+        if (map == null) return null;
+
+        String type = map.getString("type");
+
+        if (type.equals(ExpressionParser.TYPE_MAP)) {
+            JsonObject result = new JsonObject();
+
+            ReadableArray keyValues = map.getArray("value");
+            for (int i = 0; i < keyValues.size(); i++) {
+                ReadableArray keyValue = keyValues.getArray(i);
+                String key = keyValue.getMap(0).getString("value");
+
+                result.add(key, typedToJsonElement(keyValue.getMap(1)));
+            }
+            return result;
+        }
+        else if (type.equals(ExpressionParser.TYPE_ARRAY)) {
+            ReadableArray arrayValue = map.getArray("value");
+            JsonArray result = new JsonArray(arrayValue.size());
+            for (int i = 0; i < arrayValue.size(); i++) {
+                result.add(typedToJsonElement(arrayValue.getMap(i)));
+            }
+            return result;
+        }
+        else if (type.equals(ExpressionParser.TYPE_BOOL)) {
+            return new JsonPrimitive(map.getBoolean("value"));
+        }
+        else if (type.equals(ExpressionParser.TYPE_NUMBER)) {
+            return new JsonPrimitive(map.getDouble("value"));
+        }
+        else if (type.equals(ExpressionParser.TYPE_STRING)) {
+            return new JsonPrimitive(map.getString("value"));
+        }
+        else {
+            throw new RuntimeException(String.format("Unrecognized type {}", map.getString("type")));
+        }
+    }
 
     public static WritableArray toWritableArray(JsonArray array) {
         WritableArray writableArray = Arguments.createArray();