+++ android/rctmgl/src/main/java/com/mapbox/rctmgl/utils/DownloadMapImageTask.java
@@ -1,10 +1,12 @@
 package com.mapbox.rctmgl.utils;
 
 import android.content.Context;
+import android.content.res.Resources;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
 import android.os.AsyncTask;
 import android.util.Base64;
+import android.util.DisplayMetrics;
 import android.util.Log;
 
 import com.facebook.common.logging.FLog;
@@ -21,7 +23,7 @@ import static com.facebook.react.views.textinput.ReactTextInputManager.TAG;
  * Created by nickitaliano on 9/13/17.
  */
 
-public class DownloadMapImageTask extends AsyncTask<Map.Entry<String, String>, Void, List<Map.Entry<String, Bitmap>>> {
+public class DownloadMapImageTask extends AsyncTask<Map.Entry<String, ImageEntry>, Void, List<Map.Entry<String, Bitmap>>> {
     public static final String LOG_TAG = DownloadMapImageTask.class.getSimpleName();
 
     private Context mContext;
@@ -40,21 +42,24 @@ public class DownloadMapImageTask extends AsyncTask<Map.Entry<String, String>, V
 
     @SafeVarargs
     @Override
-    protected final List<Map.Entry<String, Bitmap>> doInBackground(Map.Entry<String, String>... objects) {
+    protected final List<Map.Entry<String, Bitmap>> doInBackground(Map.Entry<String, ImageEntry>... objects) {
+        Resources resources = mContext.getResources();
+        DisplayMetrics metrics = resources.getDisplayMetrics();
         List<Map.Entry<String, Bitmap>> images = new ArrayList<>();
 
-        for (Map.Entry<String, String> object : objects) {
-            String uri = object.getValue();
+        for (Map.Entry<String, ImageEntry> object : objects) {
+            ImageEntry imageEntry = object.getValue();
 
+            String uri = imageEntry.uri;
             if (uri.contains("://")) { // has scheme attempt to get bitmap from url
                 try {
-                    Bitmap bitmap = BitmapUtils.getBitmapFromURL(uri, null);
+                    Bitmap bitmap = BitmapUtils.getBitmapFromURL(uri, getBitmapOptions(metrics, imageEntry.scale));
                     images.add(new AbstractMap.SimpleEntry<String, Bitmap>(object.getKey(), bitmap));
                 } catch (Exception e) {
                     Log.w(LOG_TAG, e.getLocalizedMessage());
                 }
             } else if (uri.startsWith("/")) {
-                Bitmap bitmap = BitmapUtils.getBitmapFromPath(uri, null);
+                Bitmap bitmap = BitmapUtils.getBitmapFromPath(uri, getBitmapOptions(metrics, imageEntry.scale));
                 images.add(new AbstractMap.SimpleEntry<String, Bitmap>(object.getKey(), bitmap));
             } else if (uri.startsWith("data:image")) {
                 try {
@@ -66,7 +71,7 @@ public class DownloadMapImageTask extends AsyncTask<Map.Entry<String, String>, V
                 }
             } else {
                 // local asset required from JS require('image.png') or import icon from 'image.png' while in release mode
-                Bitmap bitmap = BitmapUtils.getBitmapFromResource(mContext, uri, null);
+                Bitmap bitmap = BitmapUtils.getBitmapFromResource(mContext, uri, getBitmapOptions(metrics, imageEntry.scale));
                 if (bitmap != null) {
                     images.add(new AbstractMap.SimpleEntry<String, Bitmap>(object.getKey(), bitmap));
                 } else {
@@ -93,5 +98,13 @@ public class DownloadMapImageTask extends AsyncTask<Map.Entry<String, String>, V
         if (mCallback != null) {
             mCallback.onAllImagesLoaded();
         }
+    }
+
+    private BitmapFactory.Options getBitmapOptions(DisplayMetrics metrics, Double scale) {
+        BitmapFactory.Options options = new BitmapFactory.Options();
+        options.inScreenDensity = metrics.densityDpi;
+        options.inTargetDensity = metrics.densityDpi;
+        options.inDensity = (int)((double)DisplayMetrics.DENSITY_DEFAULT * scale);
+        return options;
     }
 }