+++ android/rctmgl/src/main/java/com/mapbox/rctmgl/utils/ExpressionParser.java
@@ -3,7 +3,11 @@ package com.mapbox.rctmgl.utils;
 import com.facebook.react.bridge.Dynamic;
 import com.facebook.react.bridge.ReadableArray;
 import com.facebook.react.bridge.ReadableMap;
+import com.google.gson.JsonArray;
 import com.mapbox.mapboxsdk.style.expressions.Expression;
+import com.mapbox.mapboxsdk.style.expressions.Expression.Converter;
+
+import org.json.JSONArray;
 
 import java.util.Locale;
 
@@ -15,29 +19,19 @@ public class ExpressionParser {
     static final String TYPE_BOOL = "boolean";
 
     public static Expression from(ReadableArray rawExpressions) {
-        StringBuilder builder = new StringBuilder();
-
         if (rawExpressions == null || rawExpressions.size() == 0) {
             return null;
         }
 
-        builder.append("[");
-        for (int i = 0; i < rawExpressions.size(); i++) {
-            ReadableMap item = rawExpressions.getMap(i);
-
-            String curExpression = stringExpression(item);
-            if (!curExpression.isEmpty()) {
-                builder.append(curExpression);
-
-                if (i < rawExpressions.size() - 1) {
-                    builder.append(",");
-                }
-            }
-        }
-        builder.append("]");
-
-        return Expression.raw(builder.toString());
+        return Expression.Converter.convert(ConvertUtils.toJsonArray(rawExpressions));
     }
+
+    public static Expression fromTyped(ReadableMap rawExpressions) {
+        JsonArray array = (JsonArray)ConvertUtils.typedToJsonElement(rawExpressions);
+        return Expression.Converter.convert(array);
+    }
+
+
 
     public static Expression from(ReadableMap rawExpression) {
         return Expression.raw("[" + stringExpression(rawExpression) + "]");