+++ ios/RCTMGL/RCTMGLLayer.m
@@ -11,7 +11,7 @@
 #import "RCTMGLStyleValue.h"
 #import "RCTMGLUtils.h"
 #import "FilterParser.h"
-#import "FilterList.h"
+#import "RCTMGLMapView.h"
 
 @implementation RCTMGLLayer
 
@@ -42,7 +42,7 @@
     _aboveLayerID = aboveLayerID;
     if (_styleLayer != nil) {
         [self removeFromMap:_style];
-        [self insertAbove:[_style layerWithIdentifier:aboveLayerID]];
+        [self insertAbove:aboveLayerID];
     }
 }
 
@@ -55,7 +55,7 @@
     _belowLayerID = belowLayerID;
     if (_styleLayer != nil) {
         [self removeFromMap:_style];
-        [self insertBelow:[_style layerWithIdentifier:belowLayerID]];
+        [self insertBelow:_belowLayerID];
     }
 }
 
@@ -94,8 +94,9 @@
     }
 }
 
-- (void)addToMap:(MGLStyle *)style
+- (void)addToMap:(RCTMGLMapView*) map style:(MGLStyle *)style
 {
+    _map = map;
     _style = style;
     if (_id == nil) {
       RCTLogError(@"Cannot add a layer without id to the map: %@", self);
@@ -106,10 +107,11 @@
         _styleLayer = existingLayer;
     } else {
         _styleLayer = [self makeLayer:style];
-        [self insertLayer];
+        [self insertLayer: map];
     }
     
     [self addStyles];
+    [self addedToMap];
 }
 
 - (void)removeFromMap:(MGLStyle *)style
@@ -133,25 +135,31 @@
                                  userInfo:nil];
 }
 
+- (void)addedToMap
+{
+    // override if you want
+}
+
 - (void)updateFilter:(NSPredicate *)predicate
 {
     // override if you want to update the filter
 }
 
-- (void)insertLayer
+- (void)insertLayer: (RCTMGLMapView*) map
 {
     if ([_style layerWithIdentifier:_id] != nil) {
         return; // prevent layer from being added twice
     }
 
     if (_aboveLayerID != nil) {
-        [self insertAbove:[_style layerWithIdentifier:_aboveLayerID]];
+        [self insertAbove:_aboveLayerID];
     } else if (_belowLayerID != nil) {
-        [self insertBelow:[_style layerWithIdentifier:_belowLayerID]];
+        [self insertBelow:_belowLayerID];
     } else if (_layerIndex != nil) {
         [self insertAtIndex:(NSUInteger)_layerIndex];
     } else {
         [_style addLayer:_styleLayer];
+        [_map layerAdded:_styleLayer];
     }
     
     [self setZoomBounds];
@@ -168,20 +176,27 @@
     }
 }
 
--(void)insertAbove:(MGLStyleLayer*)layer
+-(void)insertAbove:(NSString*)aboveLayerIDs
 {
-    if (![self _hasInitialized]) {
-        return;
-    }
-    [_style insertLayer:_styleLayer aboveLayer:layer];
+    [_map waitForLayerWithID: aboveLayerIDs then:^void (MGLStyleLayer* layer) {
+        if (![self _hasInitialized]) {
+            return;
+        }
+        [_style insertLayer:_styleLayer aboveLayer:layer];
+        [_map layerAdded:_styleLayer];
+    }];
 }
 
--(void)insertBelow:(MGLStyleLayer*)layer
+-(void)insertBelow:(NSString*)belowLayerID
 {
-    if (![self _hasInitialized]) {
-        return;
-    }
-    [_style insertLayer:_styleLayer belowLayer:layer];
+    [_map waitForLayerWithID: belowLayerID then:^void (MGLStyleLayer* layer) {
+        if (![self _hasInitialized]) {
+            return;
+        }
+        
+        [_style insertLayer:_styleLayer belowLayer:layer];
+        [_map layerAdded:_styleLayer];
+    }];
 }
 
 -(void)insertAtIndex:(NSUInteger)index
@@ -190,6 +205,7 @@
         return;
     }
     [_style insertLayer:_styleLayer atIndex:index];
+    [_map layerAdded:_styleLayer];
 }
 
 - (void)addImage:(NSString *)url
@@ -197,7 +213,7 @@
     if (url == nil) {
         return;
     }
-    [RCTMGLUtils fetchImage:_bridge url:url callback:^(NSError *error, UIImage *image) {
+    [RCTMGLUtils fetchImage:_bridge url:url scale:1.0 callback:^(NSError *error, UIImage *image) {
         if (image != nil) {
             [_style setImage:image forName:url];
         }
@@ -206,7 +222,7 @@
 
 - (NSPredicate*)buildFilters
 {
-    return [FilterParser parse:[[FilterList alloc] initWithArray:_filter]];
+    return [FilterParser parse:_filter];
 }
 
 - (BOOL)_hasInitialized