+++ ios/RCTMGL/RCTMGLStyleValue.m
@@ -27,9 +27,18 @@
     } else if ([_styleType isEqualToString:@"vector"] && [expressionJSON respondsToSelector:@selector(objectEnumerator)] && [[[(NSArray*)expressionJSON objectEnumerator] nextObject] isKindOfClass:[NSNumber class]]) {
         CGVector vector = [RCTMGLUtils toCGVector:(NSArray<NSNumber *> *)expressionJSON];
         return [NSExpression expressionWithMGLJSONObject:[NSValue valueWithCGVector:vector]];
-    } else if ([_styleType isEqual:@"edgeinsets"] && [expressionJSON isKindOfClass:[NSNumber class]]){
+    } else if ([_styleType isEqualToString:@"image"] && [expressionJSON isKindOfClass:[NSDictionary class]]) {
+        return [NSExpression expressionForConstantValue:[self getImageURI]];
+    } else if ([_styleType isEqual:@"edgeinsets"] && [expressionJSON respondsToSelector:@selector(objectEnumerator)] && [[[(NSArray*)expressionJSON objectEnumerator] nextObject] isKindOfClass:[NSNumber class]]){
         UIEdgeInsets edgeInsets = [RCTMGLUtils toUIEdgeInsets:(NSArray<NSNumber *> *)expressionJSON];
         return [NSExpression expressionWithMGLJSONObject:[NSValue valueWithUIEdgeInsets:edgeInsets]];
+    } else if ([_styleType isEqualToString:@"enum"] && [expressionJSON isKindOfClass:[NSNumber class]]) {
+        // ensure we pass through values as NSUInteger when mapping to an MGL enum
+        NSUInteger uintValue = [(NSNumber*)expressionJSON unsignedIntegerValue];
+        id rawValue = [NSValue value:&uintValue withObjCType:@encode(NSUInteger)];
+        return [NSExpression expressionWithMGLJSONObject:rawValue];
+    } else if ([expressionJSON respondsToSelector:@selector(objectEnumerator)] && [[[(NSArray*)expressionJSON objectEnumerator] nextObject] isKindOfClass:[NSNumber class]]) {
+        return [NSExpression expressionForConstantValue:expressionJSON];
     } else {
         return [NSExpression expressionWithMGLJSONObject:expressionJSON];
     }
@@ -76,15 +85,36 @@
     return object;
 }
 
+
+
 - (BOOL)shouldAddImage
 {
-    NSString *imageURI = (NSString *)expressionJSON;
+    NSString *imageURI = [self getImageURI];
+    
     return [imageURI containsString:@"://"];
 }
 
 - (NSString *)getImageURI
 {
-    return (NSString *)expressionJSON;
+    if ([expressionJSON isKindOfClass:[NSDictionary class]]) {
+        return [expressionJSON valueForKey:@"uri"];
+    } else {
+        return (NSString *)expressionJSON;
+    }
+}
+
+- (double)getImageScale
+{
+    if ([expressionJSON isKindOfClass:[NSDictionary class]]) {
+        id scale = [expressionJSON valueForKey:@"scale"];
+        if ([scale isKindOfClass:[NSNumber class]]) {
+            return [scale doubleValue];
+        } else {
+            return 1.0;
+        }
+    } else {
+        return 1.0;
+    }
 }
 
 - (MGLTransition)getTransition