+++ CHANGELOG.md
@@ -2,8 +2,31 @@
 
 ### Breaking changes:
 
-* isTelemetryEnabled removed (as no longer supported on android) [#1](https://github.com/mfazekas/maps/pull/1)
-* Camera related properties on MapView now have to be specified on a camera object:
+* iOS mapbox libraries updated to [5.0.0](https://github.com/mapbox/mapbox-gl-native/releases/tag/ios-v5.0.0) android libraries updated to [8.0.0](https://github.com/mapbox/mapbox-gl-native/releases/tag/android-v8.0.0)
+* `StyleSheet.create` removed.  
+Mapbox styles are now just a map no need for `StyleSheet.create`.  
+`StylesSheet.identity` also removed, use expressions array instead:
+   ```jsx
+   mapboxStyle=MapboxGL.Stylesheet.create({..., fillColor: MapboxGL.Stylesheet.identity('color') ...})
+   ...
+   <MapView
+     ...
+     <FillLayer style={mapboxStyle}... />
+   </MapView>
+   ```
+   
+   is now:
+   ```jsx
+   mapboxStyle={..., fillColor: ['get', 'color'] ...}
+   ...
+   <MapView
+     ...
+     <FillLayer style={mapboxStyle}... />
+   </MapView>
+   ```
+  See [docs/StyleSheet.md](docs/StyleSheet.md) for more examples
+* `isTelemetryEnabled` removed (as no longer supported on android) [#1](https://github.com/mfazekas/maps/pull/1)
+* Camera related properties on `MapView` now have to be specified on a camera object:
    ```jsx
    <MapView
       zoomLevel={8}
@@ -14,7 +37,7 @@
    </MapView>
    ```
     
-   is now
+   is now:
    
    ```jsx
    <MapView
@@ -26,7 +49,8 @@
      />
    </MapView>
    ```
-* User tracking properties moved from MapView to Camera
+   See [docs/Camera.md](docs/Camera.md) for more examples
+* User tracking properties moved from `MapView` to `Camera`
    ```jsx
    <MapView
       userTrackingMode={UserTrackingModes.Follow}
@@ -36,7 +60,7 @@
    </MapView>
    ```
     
-   is now
+   is now:
    
    ```jsx
    <MapView
@@ -44,7 +68,7 @@
    >
      <Camera
         followUserLocation=true
-        followUserMode="follow"
+        followUserMode="normal"
      />
    </MapView>
    ```